refactoring to enable unit tests of js helper functions.
